Buoi7/library_app/requirements.txt
connexion >= 2.6.0, < 3.0.0
connexion[swagger-ui] >= 2.6.0, < 3.0.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
flask < 2.3.0
werkzeug < 2.3.0
flask-pymongo < 3.0.0
pymongo == 4.16.0
python-dotenv == 1.2.2
